pyproject.toml
[project]
name = "Reasoning-Engine-CoT"
version = "0.0.1"
description = "Chain-of-Thought fine-tuning specialist for logic and math puzzles."
authors = [{ name = "Egemen Çankaya", email = "egemencankaya14@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
dependencies = [
    "torch>=2.3.0",
    "transformers>=4.41.0",
    "datasets>=2.19.0",
    "peft>=0.11.0",
    "bitsandbytes>=0.43.0; platform_system != 'Windows'",
    "bitsandbytes-windows>=0.0.1; platform_system == 'Windows'",
    "trl>=0.9.0",
    "unsloth>=2024.7.0",
    "accelerate>=0.30.0",
    "streamlit>=1.32.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "pyyaml>=6.0.1",
    "huggingface_hub<1.0",
    "psutil>=5.9.0",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://EgeCankaya.github.io/Reasoning-Engine-CoT/"
Repository = "https://github.com/EgeCankaya/Reasoning-Engine-CoT"
Documentation = "https://EgeCankaya.github.io/Reasoning-Engine-CoT/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-cov>=5.0.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "mypy>=0.991",
    
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reasoning_engine_cot"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true


